refactor: extract slack payload construction into slackz.js
